+++ util/bytemanipulation.go
@@ -14,6 +14,13 @@ func ReverseBytes(a []byte) []byte {
 	return tmp
 }
 
+// ReverseBytesInPlace reverses the bytes (little endian/big endian) in place (no extra memory allocation).
+func ReverseBytesInPlace(a []byte) {
+	for i, j := 0, len(a)-1; i < j; i, j = i+1, j-1 {
+		a[i], a[j] = a[j], a[i]
+	}
+}
+
 // LittleEndianBytes returns a byte array in little endian from an unsigned integer of 32 bytes.
 func LittleEndianBytes(v uint32, l uint32) []byte {
 	buf := make([]byte, 4)